Deal with 404 errors for turtle
Before this change, a request to a nonexistent turtle page (eg
/foobar?_representation=turtle) would never respond, because
TurtleRepresentation#toResponse() returned a null rather than a valid
promise of a response.

However it didn't feel right to serve error pages in turtle format.

I've therefore changed the Representation interface to remove the
toResponse() method, which in practice was only used for errors (and
also the 202 Accepted response), and never served an actual record or
list of records.

This hardcodes errors to always be HTML for the query frontend, and JSON
for the create record API.