Ensure fields register starts successfully
Currently, a change to this repository causes all registers to be
redeployed to heroku simultaneously.  This can cause the field register
to throw a java.util.concurrent.TimeoutException during init() when
calling the register register.

This change forces the FieldRegister class to swallow all exceptions
during initialization, because if the field register fails to start,
then it will cause all other registers to fail to initialize too.