Tell google which URL for a resource is canonical
This adds a Link: header to tell google and other crawlers which url is
the canonical URL for a resource.  RFC5988 defines the Link: header, and
https://support.google.com/webmasters/answer/139066?hl=en gives a good
summary of using it to define canonical URLs.

This makes it so that:

    /hash/abcd.json

sets a header

    Link: </hash/abcd> rel="canonical"

and

    /key1/value1.json

sets a header

    Link: </key1/value1> rel="canonical"

This works for all representations as it's done at the Rest level.