Correctly import fields with cardinality of "many"
In openregister/field.register@e203c77a5f83eb7582006bc30e987735db5ca6a0
I updated fields to use a cardinality of "n" and remove usages of the
"string-set" datatype.  However that [broke the public body register][1].

The import process doesn't understand cardinality, so just imports
things as strings where we expect them to be json arrays.  This commit
changes the import process to import fields with a cardinality of many
to json array values.  The separator for array values is the semicolon.

[1]: http://public-body.openregister.org/public-body/uk-trade-investment.ttl