Make alpha work with beta public-body register
We've struggled to move the register, field and public-body registers to
the beta product, because is breaks all the other alpha registers such
as products-of-animal-origin-premises.

It turns out this was because when the alpha register tries to fetch the
public-body registry for the given register, it uses the
/public-body/foo?_representation=json URL which the beta no longer
supports.  The /public-body/foo.json URL is supported by both alpha and
beta, so this change is safe to merge in isolation.  Once this is
merged, we should be able to migrate the register, field and public-body
registers to the beta product.